pyproject.toml
[tool.poetry]
name = "django-fritzconnection"
version = "0.0.1-alpha"
description = "Web based FritzBox management using Python/Django."
authors = ["JensDiemer <git@jensdiemer.de>"]
homepage = "https://github.com/jedie/django-fritzconnection"
packages = [
    { include = "djfritz" },
    { include = "djfritz_project" },
]
keywords=['fritzbox','django', 'fritzconnect']
classifiers = [
    # http://pypi.python.org/pypi?%3Aaction=list_classifiers
    'Development Status :: 1 - Planning',
#    'Development Status :: 2 - Pre-Alpha',
#    'Development Status :: 3 - Alpha',
#    'Development Status :: 4 - Beta',
#    'Development Status :: 5 - Production/Stable',
    'Environment :: Web Environment',
    'Intended Audience :: End Users/Desktop',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3 :: Only',
    'Framework :: Django',
    'Topic :: Database :: Front-Ends',
    'Topic :: Documentation',
    'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
    'Topic :: Internet :: WWW/HTTP :: Site Management',
    'Topic :: Internet :: WWW/HTTP :: WSGI :: Application',
    'Operating System :: OS Independent',
]
include = ['AUTHORS', 'LICENSE', 'README.md']
readme='README.md'

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/jedie/django-fritzconnection/issues"


[tool.poetry.dependencies]
python = ">=3.8,<4.0.0"
django = "*"
fritzconnection = "*"  # https://github.com/kbr/fritzconnection
colorlog = "*"  # https://github.com/borntyping/python-colorlog
django-debug-toolbar = "*"  # http://django-debug-toolbar.readthedocs.io/en/stable/changes.html
django-tools = "*"  # https://github.com/jedie/django-tools/
bx_py_utils = "*"  # https://github.com/boxine/bx_py_utils
bx_django_utils = "*"  # https://github.com/boxine/bx_django_utils
django-reversion-compare = "*"  # https://github.com/jedie/django-reversion-compare/
django-tagulous = "*"  # https://github.com/radiac/django-tagulous
django-admin-sortable2 = "*"  # https://github.com/jrief/django-admin-sortable2

[tool.poetry.dev-dependencies]
dev_shell = "*"  # https://github.com/jedie/dev-shell
poetry-publish = "*"  # https://github.com/jedie/poetry-publish
cmd2_ext_test = "*"
tox = "*"
pytest = "*"
pytest-randomly = "*"
pytest-cov = "*"
pytest-django = "*"
pytest-darker = "*"  # https://github.com/akaihola/pytest-darker
pytest-flake8 = "*"
pytest-isort = "*"
# TODO: pytest-mypy = "*"
coveralls = "*"
isort = "*"
flake8 = "*"
flynt = "*"
darker = "*"  # https://github.com/akaihola/darker
pyupgrade = "*"
model_bakery = "*"  # https://github.com/model-bakers/model_bakery
beautifulsoup4 = "*"
lxml = "*"


[tool.poetry.scripts]
devshell = 'djfritz_project.dev_shell:devshell_cmdloop'
run_testserver = 'djfritz_project.manage:start_test_server'


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


[tool.darker]
src = ['.']
revision = "origin/main..."
line_length = 100
verbose = true
skip_string_normalization = true
diff = false
check = false
stdout = false
isort = true
lint = [
    "flake8",
]
log_level = "INFO"


[tool.flynt]
line_length = 100


[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/config_files/#pyprojecttoml-preferred-format
atomic=true
profile='black'
line_length=100
skip_glob=[".*", "*/htmlcov/*","*/migrations/*"]
known_first_party=["djfritz","djfritz_project","djfritz_tests"]
lines_after_imports=2


[tool.coverage.run]
omit = [".*"]


[tool.pytest.ini_options]
# https://docs.pytest.org/en/latest/customize.html#pyproject-toml
minversion = "6.0"
DJANGO_SETTINGS_MODULE="djfritz_project.settings.tests"

# Don't overwrite settings.DEBUG:
django_debug_mode="keep"

norecursedirs = ".* .git __pycache__ coverage* dist htmlcov migrations"
# sometimes helpfull "addopts" arguments:
#    -vv
#    --verbose
#    --capture=no
#    --trace-config
#    --full-trace
#    -p no:warnings
addopts = """
    --reuse-db
    --nomigrations
    --cov=.
    --cov-report term-missing
    --cov-report html
    --cov-report xml
    --no-cov-on-fail
    --showlocals
    --darker
    --flake8
    --isort
    --doctest-modules
    --failed-first
    --last-failed-no-failures all
    --new-first
    -p no:randomly
"""
# TODO: --mypy


[tool.tox]
# https://tox.readthedocs.io/en/latest/example/basic.html#pyproject-toml-tox-legacy-ini
legacy_tox_ini = """
[tox]
isolated_build = True
envlist = px310,py39,py38
skip_missing_interpreters = True

[testenv]
passenv = *
whitelist_externals = pytest
commands =
    pytest djfritz djfritz_project
"""
